requirements.txt
PyQt6
pycryptodome
argon2-cffi
pyinstaller